Nest generator under module
